+++ src/imio/events/core/browser/actions.py
@@ -1,3 +1,4 @@
+from imio.events.core.contents import IAgenda
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone import api
 from plone.app.content.browser.actions import (
@@ -15,9 +16,16 @@ class DeleteConfirmationForm(BaseDeleteConfirmationForm):
     def items_to_delete(self):
         count = len(self.context.items())
         if count >= 1:
+            container_to_delete = (
+                _("Agenda") if IAgenda.providedBy(self.context) else _("Folder")
+            )
             txt = _(
-                "Agenda ${val1} can't be removed because it contains ${val2} event(s)",
-                mapping={"val1": self.context.Title(), "val2": count},
+                "${val1} ${val2} can't be removed because it contains ${val3} element(s)",
+                mapping={
+                    "val1": container_to_delete,
+                    "val2": self.context.Title(),
+                    "val3": count,
+                },
             )
             msg = translate(txt, context=self.request)
             api.portal.show_message(msg, self.request, type="warn")
@@ -31,9 +39,16 @@ class DeleteActionView(BaseDeleteActionView):
     def action(self, obj):
         count = len(obj.items())
         if count >= 1:
+            container_to_delete = (
+                _("Agenda") if IAgenda.providedBy(self.context) else _("Folder")
+            )
             txt = _(
-                "Agenda ${val1} can't be removed because it contains ${val2} event(s)",
-                mapping={"val1": obj.Title(), "val2": count},
+                "${val1} ${val2} can't be removed because it contains ${val3} element(s)",
+                mapping={
+                    "val1": container_to_delete,
+                    "val2": obj.Title(),
+                    "val3": count,
+                },
             )
             msg = translate(txt, context=self.request)
             self.errors.append(msg)